backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
requests==2.31.0
openai==1.58.1
pydantic==2.6.4
email-validator==2.1.0.post1
typing-extensions>=4.11.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
python-dotenv==1.0.0
passlib[bcrypt]==1.7.4